updating psot form and comment form for file uploads

+++ library/taxonomy/file/file.taxonomy.php
@@ -126,15 +126,15 @@ class FileTaxonomy extends Entity {
     {
         if ( $this->hasError ) return $this;
 
-        // if ( $this->exists() === false ) return $this->error(e()->file_not_exists);
-        // if ( file_exists($this->path) === false ) return $this->error(e()->file_not_exists);
-
+        if ( $this->exists() === false ) return $this->error(e()->file_not_exists_in_db);
         if ( file_exists($this->path) === false ) {
-            // ... Ignore, since it is already deleted in the disk.
-        } else {
-            $re = @unlink($this->path);
-            if ( $re === false ) return $this->error(e()->file_delete_failed);
+            // the file exists in wc_files table, but actual file does not exists, so, it will delete the record.
+            parent::delete();
+            return $this->error(e()->file_not_exists_in_disk);
         }
+
+        $re = @unlink($this->path);
+        if ( $re === false ) return $this->error(e()->file_delete_failed);
 
         return parent::delete();
     }
